requirements.txt
metaapi-cloud-sdk
python-dotenv
dash
dash-bootstrap-components
TA-Lib
pandas
pandas_ta